se arreglo tipos int/float -> int

+++ suma.h
@@ -1,8 +1,8 @@
 #ifndef
 #define suma_h
 
-float suma(int a, int b){
-    float c=a+b;
+int suma(int a, int b){
+    int c=a+b;
     return c;
 }
